Patch the infinity coin glitch
It happens when you open a game from your igloo and keep getting infinity coins when you end the game. When doing it from a game, I noticed it kept going to a room 901 instead of an igloo room id so I just checked if it's equal. If anyone tries to end the game and tries to keep getting the coins, they will get stuck.